fix get resources

+++ neural_style_transfer/get_resources.py
@@ -3,7 +3,8 @@ import zipfile
 from urllib import request
 
 def download_and_extract(url, fname):
-    request.urlretrieve(url, fname)
+    if not os.path.exists(fname):
+        request.urlretrieve(url, fname)
 
     if not os.path.exists(fname.split('.')[0]):
         with zipfile.ZipFile(fname, 'r') as f:
